backend/app/requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
python-dotenv
google-generativeai
faiss-cpu
tqdm
pymupdf
pydantic
requests
